[ADF-4302] - move cloud folder from app-layout to component
