+++ demo-shell/src/app/app.routes.ts
@@ -39,16 +39,16 @@ import { DemoPermissionComponent } from './components/permissions/demo-permissio
 import { ReportIssueComponent } from './components/report-issue/report-issue.component';
 import { AppComponent } from './app.component';
 import { TreeViewSampleComponent } from './components/tree-view/tree-view-sample.component';
-import { CloudLayoutComponent } from './components/app-layout/cloud/cloud-layout.component';
-import { ProcessesCloudDemoComponent } from './components/app-layout/cloud/processes-cloud-demo.component';
-import { TaskDetailsCloudDemoComponent } from './components/app-layout/cloud/task-details-cloud-demo.component';
-import { AppsCloudDemoComponent } from './components/app-layout/cloud/apps-cloud-demo.component';
-import { TasksCloudDemoComponent } from './components/app-layout/cloud/tasks-cloud-demo.component';
-import { StartTaskCloudDemoComponent } from './components/app-layout/cloud/start-task-cloud-demo.component';
-import { StartProcessCloudDemoComponent } from './components/app-layout/cloud/start-process-cloud-demo.component';
+import { AppsCloudDemoComponent } from './components/cloud/apps-cloud-demo.component';
+import { PeopleGroupCloudDemoComponent } from './components/cloud/people-groups-cloud-demo.component';
+import { CloudLayoutComponent } from './components/cloud/cloud-layout.component';
+import { TasksCloudDemoComponent } from './components/cloud/tasks-cloud-demo.component';
+import { ProcessesCloudDemoComponent } from './components/cloud/processes-cloud-demo.component';
+import { StartTaskCloudDemoComponent } from './components/cloud/start-task-cloud-demo.component';
+import { StartProcessCloudDemoComponent } from './components/cloud/start-process-cloud-demo.component';
+import { TaskDetailsCloudDemoComponent } from './components/cloud/task-details-cloud-demo.component';
+import { ProcessDetailsCloudDemoComponent } from './components/cloud/process-details-cloud-demo.component';
 import { TemplateDemoComponent } from './components/template-list/template-demo.component';
-import { PeopleGroupCloudDemoComponent } from './components/app-layout/cloud/people-groups-cloud-demo.component';
-import { ProcessDetailsCloudDemoComponent } from './components/app-layout/cloud/process-details-cloud-demo.component';
 
 export const appRoutes: Routes = [
     { path: 'login', component: LoginComponent },
