+++ demo-shell/src/app/components/cloud/apps-cloud-demo.component.ts
