+++ demo-shell/src/app/components/cloud/cloud-filters-demo.component.scss
