+++ demo-shell/src/app/components/cloud/cloud-layout.component.scss
