+++ demo-shell/src/app/components/cloud/cloud-settings.component.ts
