+++ demo-shell/src/app/components/cloud/people-groups-cloud-demo.component.scss
