+++ demo-shell/src/app/components/cloud/process-details-cloud-demo.component.scss
