+++ demo-shell/src/app/components/cloud/services/cloud-layout.service.ts
