+++ demo-shell/src/app/components/cloud/task-details-cloud-demo.component.scss
